+++ mlos_bench/mlos_bench/tests/environments/local/composite_local_env_test.py
@@ -9,7 +9,8 @@ import sys
 from datetime import datetime, timedelta
 
 from mlos_bench.tunables.tunable_groups import TunableGroups
-from mlos_bench.tests.environments.local import create_composite_local_env, check_env_success
+from mlos_bench.tests.environments import check_env_success
+from mlos_bench.tests.environments.local import create_composite_local_env
 
 
 def test_composite_env(tunable_groups: TunableGroups) -> None:
