+++ mlos_bench/mlos_bench/tests/environments/local/local_env_test.py
@@ -8,7 +8,8 @@ Unit tests for LocalEnv benchmark environment.
 import pytest
 
 from mlos_bench.tunables.tunable_groups import TunableGroups
-from mlos_bench.tests.environments.local import create_local_env, check_env_success
+from mlos_bench.tests.environments import check_env_success
+from mlos_bench.tests.environments.local import create_local_env
 
 
 def test_local_env(tunable_groups: TunableGroups) -> None:
