+++ mlos_bench/mlos_bench/tests/environments/local/local_env_vars_test.py
@@ -10,7 +10,8 @@ import sys
 import pytest
 
 from mlos_bench.tunables.tunable_groups import TunableGroups
-from mlos_bench.tests.environments.local import create_local_env, check_env_success
+from mlos_bench.tests.environments import check_env_success
+from mlos_bench.tests.environments.local import create_local_env
 
 
 def _run_local_env(tunable_groups: TunableGroups, shell_subcmd: str, expected: dict) -> None:
